requirements.txt
debugpy==1.6.0
entrypoints==0.4
matplotlib==3.6.2
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.22.3
parso==0.8.3
Pillow==9.3.0
platformdirs==2.5.1
pure-eval==0.2.2
python-dotenv==0.21.0
requests==2.28.1
requests-toolbelt==0.9.1
scikit-learn==1.1.2
scipy==1.9.0
sklearn==0.0
stack-data==0.2.0
urllib3==1.26.9
virtualenv==20.13.4
wcwidth==0.2.5
Flask-Reuploaded==1.2.0
Flask-Uploads==0.1.3
Flask-WTF==1.0.1
gunicorn==20.1.0
Werkzeug==2.2.2
detecto
pyngrok==4.1.1
ngrok
